Refactors code, fixes changed packages
