+++ library/src/main/java/pl/netigen/rateus/RateUs.java
@@ -8,7 +8,7 @@ import androidx.annotation.StringRes;
 import androidx.appcompat.app.AppCompatActivity;
 
 import pl.netigen.netigenapi.R;
-import pl.netigen.netigenapi.Utils;
+import pl.netigen.utils.Utils;
 
 import static android.content.Context.MODE_PRIVATE;
